Add run_example.sh for simple_app/, as a single script to run the entire setup
This commit introduces run_example.sh as a single wrapper script to package
the multiple steps, from instructions.md, to build and setup the simple_app
example program. The objective of this kind of packaging is to simplify the
on-boarding for new users who wish to experiment with the Certifier Library
and our example programs.

Usage: run_example.sh [--help | --list] [ setup | run_test ]

By default, with no arguments, the script will run through all the steps
to compile, build and setup the app-as-server, app-as-client invocation
to demo that the Certifier library is basically working.

--help will give help / usage info.

--list will display the functions implementing each step.
User can [laboriously] invoke this script for each step, using this list, as:

  $ run_example.sh show_env
  $ run_example.sh build_utilities
  $ run_example.sh gen_policy_and_self_signed_cert
  [...]
  $ run_example.sh run_app_as_client_talk_to_Cert_Service
  $ run_example.sh run_app_as_server_talk_to_Cert_Service
  $ run_example.sh run_app_to_app_trusted_service_communication

- Add run_example.md as a quick-start user's guide to this script.
- Fix build_simple_server() to also 'make dummy' in graminelib
- Provide cleanup.sh script to clean-up stale left-over processes

Execute simple_app/run_example.sh as part of CI-workflows. Fix up build.yml
to ensure that dependent s/w needed by simple_app are pre-installed.