Add full keystone support (API, certifier_service) using a software emulated environment (the shim).
This commit completes Certifier support for Keystone using the provided shim-support.
This is suitable for CI tests.  With real hardware, the shim will be replaced by
replacing "keystone_shim.cc" "keystone_api.cc" in the build.

This checkin includes all required Certifier API and certifier_service support.
In addition, it includes a "simple_example_under_keystone" in sample_apps.

build.yml changes and CI-integration to test Keystone example program with
shims will be in a follow-on commit.