(#119) Replace CCA/cca term with islet in core APIs and in code
This commit renames stuff like 'cca', 'CCA' in various Islet
APIs and interfaces, and code-dir / artifact-names to become
'islet' or ISLET, as appropriate. This naming is inline with
what was done for Keystone-based simple-app.

There are just a couple of remaining instances of cca, only
used for shim-produced 'bins', in the driver shell script.

Deleted file src/certifier.proto had crept back in. Delete it again.