Fix Go-compilation errors so 'go build simpleserver.go' succeeds.
This commit applies the new PolicyPool{} mgmt interfaces
to:
 - FilterKeystonePolicy(), ValidateKeystoneEvidence()
 - FilterIsletPolicy(), ValidateIsletEvidence()

Applies a few other fixes to method prototypes that had changed
in the original filter_policy branch.

This makes 'go build simpleserver.go' to now succeed.

But there are still run-time issues when starting the Certifier
Service. Need to be debugged separately.