(#147) Fix *.mak-files to support parallel make using -j threads
This commit cleans-up several Makefiles to now support use of
parallel make with -j <threads> . The main bug was in the
definition of the dependency between:

certifier.pb.h: certifier.pb.cc
certifier.pb.cc: certifier.proto

The former was missing in some mak-files or was mis-specified
along with the latter. Both caused parallel make to often fail.

This commit fixes this for all (commonly-used) *.mak files.

Additionally, run_example.sh is enhanced to supply -j flag,
for # of CPUs on the machine.

On CI-Build machines with 2 CPUs, the e2e run-time of all
tests drops from ~29+ mins (w/o this fix) to about 18 - 22~ mins
with this parallel make support.