Multidomain support (#162)
This change enables applications to certify themselves in multiple domains represented by multiple policy keys easily.

There are improvements in the policy store (so a single program can have many policy stores) and some refactored code to share common code for certification. There are also more extensive error messages, an improved print_packaged_claims and an example application ("multidomain-app") that illustrates certification and application interaction in secondary domains. There are two small changes to existing APIs: cc_trust_data::certify_me which no longer requires arguments and cc_trust_data::cold_init which takes more arguments. Existing example applications and tests were changed to use the new interfaces. An important feature of this design is that certifying in multiple domains does not require a mutually trusted service. Each domain is still in exclusive control of its policy key (for certification) and policy rules. This means that domains need not delegate trust decisions to keys they do not control and yet any domain can decide to certify a program in another domain and cooperate with programs so certified. So there are no trusted third parties and deployers need NEVER depend on other parties to make or enforce trust decisions (Although they can in the very unlikely case they want to). This is a core promise of Confidential Computing which we want to uphold.

Co-authored-by: Ye Li <yel@vmware.com>