(#160) Standardize & appliy clang-format rules to Certifier code base.
Starting from a near-empty .clang-format file, containing just
'Language: Cpp' and 'BasedOnStyle: Google', this change set brings-in
various clang-format'ting rules, to standardize the coding style,
indentation and formatting for the Certifier code base.

Most of the basic Google-style (such as use of 2-spaces for indentations
v/s tabs etc.) have been retained. Some changes are to do with readability,
avoiding single-line function calls / loops, and alignment of parameters in
function declarations, definitions and function calls. We have standardized
on clang-format-11 tool, which is available on both Linux and macOSX.
Currently, code in third_party/ sub-dir are not being processed.

The main files to note are: .clang-format, added to top-level root
directory. CI/scripts/check-srcfmt.sh is a stand-alone script that uses
the clang-format tool to verify that all files in the code base conform to
the style(s) chosen. 'fixall' argument can be used to fix the formatting
in any active files during development.

This script is run at the start of the CI build-job and will fail the job
if any formatting errors are found.