simple_app_python: Complete client-server secure communication channel
This commit completes the work remaining (post SHA 568aef778de389)
to setup a secure communication channel between client and server
portions of simple_app_python. This handling is modelled on the
lines of the mTLS communication pytest (SHA 17a4f15).

As Python SSL interfaces only deal with certificates / keys from
files, we do the following in this work:

- Add external interface to cc_trust_manager{} to write out the
  private key to a file.

- As part of 'get-certified' operation, the certificate and
  private-key for the client & server get written out to files
  in their respective application data-dir. The CA-root policy
  cert will also get written out.

- Introduce server_dispatch() and client_dispatch() methods, to
  run under the "app-as-server" and "app-as-client" targets.
  These methods will establish the secure SSL channel communication,
  using the certificate & private-key files persisted above.

With this work, simple_app in Python is fully functional demonstrating
the exchange of a simple 'Hello' message. (Caveat is that we
expect for a real-app, any data persisted will be written to
secure storage within a TEE.)

Apply review feedback:

- Create only one set of cert/pvt-key files for server in provisioning/ dir.

Other:
- Check for NULL from fopen(), while writing out private-key.
  Enhance test-case to exercise this negative test logic.
- Fix test.sh to regen 'sharedlib' while running test-run_example-simple_app_python target.
- Use os.path.join() to produce full path name of output file
- Delete client-/server-specific cert / key file name symbols
- Misc cleanup, to test.sh to suppress outputs from 'pushd' etc.

Minor cleanup.